refactor(changelog): tokenize subject in plugin, drop component regex
